split ResultTable into component parts
