+++ src/exampleTypes/ResultTable/index.stories.js
@@ -1,9 +1,9 @@
 import _ from 'lodash/fp'
 import React from 'react'
 import { storiesOf } from '@storybook/react'
-import TestTree from './stories/testTree'
-import ThemePicker from '../stories/themePicker'
-import { ResultTable } from '.'
+import TestTree from '../stories/testTree'
+import ThemePicker from '../../stories/themePicker'
+import ResultTable from '.'
 
 storiesOf('ExampleTypes|ResultTable', module)
   .addDecorator(ThemePicker('greyVest'))
